requirements.txt
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.1
django-cors-headers==3.7.0
djangorestframework==3.12.4
drf-yasg2==1.19.4
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
isort==5.8.0
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pylint==2.8.2
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
